Externalize DSpaceKernelInitializer so that it can be used by Integration Tests to start Kernel
